Move internal classes to Internal namespace
